pyproject.toml
[project]
name = "stageflow"
version = "0.1.0"
description = "A declarative multi-stage validation framework that provides reusable, non-mutating evaluation of data-bearing elements within multi-stage processes"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "StageFlow Team", email = "team@stageflow.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance"
]
dependencies = [
    "pydantic>=2.0.0",
    "ruamel.yaml>=0.18.0",
    "click>=8.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pylint>=3.0.0",
    "hypothesis>=6.0.0",
    "pytest-cov>=4.0.0"
]
visualization = [
    "graphviz>=0.20.0",
    "rich>=13.0.0"
]

[project.scripts]
stageflow = "stageflow.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pylint>=3.0.0",
    "hypothesis>=6.0.0",
    "pytest-cov>=4.0.0"
]

# Note: UV scripts will be available in a future version
# For now, use commands directly:
# uv run pytest
# uv run ruff check .
# uv run ruff check --fix .
# uv run pylint stageflow
# uv run ruff format .
# uv run pytest --cov=stageflow --cov-report=term-missing

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--color=yes"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]

[tool.pylint.messages_control]
disable = [
    "too-few-public-methods",
    "too-many-arguments",
    "line-too-long"  # handled by ruff
]

[tool.coverage.run]
source = ["stageflow"]
omit = [
    "stageflow/examples/*",
    "tests/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]
